refactor: 채팅방 로비 위젯 생성 (#88)
