+++ src/app/chat/page.tsx
@@ -1,14 +1,6 @@
-import { getChatRoomList } from '@/entities/chat-room/api/api'
-import ChatRoomSortButton from '@/features/chat-room-sort/ui/ChatRoomSortButton'
-import { chatKeys } from '@/shared/api/query-keys'
-import { ChatRoomList } from '@/widgets/chat-room-list'
-import {
-  dehydrate,
-  HydrationBoundary,
-  QueryClient,
-} from '@tanstack/react-query'
+import { ChatRoomLobby } from '@/widgets/chat-room-lobby/ui'
+
 import Image from 'next/image'
-import { Suspense } from 'react'
 
 async function Chat({
   searchParams,
@@ -17,13 +9,6 @@ async function Chat({
 }) {
   const searchParam = await searchParams
   const sort = searchParam.sort ?? 'desc'
-  const queryClient = new QueryClient()
-
-  await queryClient.prefetchQuery({
-    queryKey: chatKeys.roomList(sort),
-    queryFn: () => getChatRoomList(sort),
-    staleTime: 60 * 1000,
-  })
 
   return (
     <div className="mt-14 mb-35.5">
@@ -38,28 +23,8 @@ async function Chat({
           sizes="100vw"
         />
       </section>
-      {/* 채팅방 목록 */}
-      <section className="mx-auto max-w-300 px-4">
-        <h1 className="text-brand-black text-3xl font-black">실시간 채팅방</h1>
-        <Suspense
-          fallback={
-            <div className="h-96 w-full animate-pulse rounded-xl bg-gray-50"></div>
-          }
-        >
-          <div className="border-b-brand-gray-100 mt-8 mb-4 border-b">
-            <div className="flex items-end justify-between">
-              <div className="text-brand-black relative pb-4 text-lg font-bold">
-                <span>전체</span>
-                <div className="bg-brand-black absolute right-0 bottom-0 left-0 h-1" />
-              </div>
-              <ChatRoomSortButton />
-            </div>
-          </div>
-          <HydrationBoundary state={dehydrate(queryClient)}>
-            <ChatRoomList />
-          </HydrationBoundary>
-        </Suspense>
-      </section>
+      {/* 채팅방 목록 영역 (필터링 메뉴, 목록) */}
+      <ChatRoomLobby sort={sort} />
     </div>
   )
 }
