--- src/widgets/chat-room-list/index.ts
@@ -1,3 +1,0 @@
-import ChatRoomList from '@/widgets/chat-room-list/ui/ChatRoomList'
-
-export { ChatRoomList }
