+++ src/widgets/chat-room-lobby/lib/formatter.ts
