+++ src/widgets/chat-room-lobby/ui/index.ts
@@ -1,0 +1,3 @@
+import ChatRoomLobby from '@/widgets/chat-room-lobby/ui/ChatRoomLobby'
+
+export { ChatRoomLobby }
